server/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0